Enforce GetTrustBundle trait
