+++ edgelet/edgelet-http-mgmt/src/server/module/delete.rs
@@ -65,6 +65,7 @@ mod tests {
 
     use edgelet_core::{MakeModuleRuntime, ModuleRuntimeState, ModuleStatus};
     use edgelet_http::route::Parameters;
+    use edgelet_test_utils::crypto::TestHsm;
     use edgelet_test_utils::module::*;
 
     use super::*;
@@ -83,11 +84,14 @@ mod tests {
         let config = TestConfig::new("microsoft/test-image".to_string());
         let module: TestModule<Error, _> =
             TestModule::new("test-module".to_string(), config, Ok(state));
-        let runtime =
-            TestRuntime::make_runtime(TestSettings::new(), TestProvisioningResult::new(), ())
-                .wait()
-                .unwrap()
-                .with_module(Ok(module));
+        let runtime = TestRuntime::make_runtime(
+            TestSettings::new(),
+            TestProvisioningResult::new(),
+            TestHsm::default(),
+        )
+        .wait()
+        .unwrap()
+        .with_module(Ok(module));
         let handler = DeleteModule::new(runtime);
         let parameters =
             Parameters::with_captures(vec![(Some("name".to_string()), "test".to_string())]);
@@ -115,11 +119,14 @@ mod tests {
         let config = TestConfig::new("microsoft/test-image".to_string());
         let module: TestModule<Error, _> =
             TestModule::new("test-module".to_string(), config, Ok(state));
-        let runtime =
-            TestRuntime::make_runtime(TestSettings::new(), TestProvisioningResult::new(), ())
-                .wait()
-                .unwrap()
-                .with_module(Ok(module));
+        let runtime = TestRuntime::make_runtime(
+            TestSettings::new(),
+            TestProvisioningResult::new(),
+            TestHsm::default(),
+        )
+        .wait()
+        .unwrap()
+        .with_module(Ok(module));
         let handler = DeleteModule::new(runtime);
         let request = Request::delete("http://localhost/modules/test")
             .body(Body::default())
