+++ edgelet/edgelet-http-mgmt/src/server/module/list.rs
@@ -101,6 +101,7 @@ mod tests {
     use chrono::prelude::*;
     use edgelet_core::{MakeModuleRuntime, ModuleRuntimeState, ModuleStatus};
     use edgelet_http::route::Parameters;
+    use edgelet_test_utils::crypto::TestHsm;
     use edgelet_test_utils::module::*;
     use futures::Stream;
     use management::models::ModuleList;
@@ -121,11 +122,14 @@ mod tests {
         let config = TestConfig::new("microsoft/test-image".to_string());
         let module: TestModule<Error, _> =
             TestModule::new("test-module".to_string(), config, Ok(state));
-        let runtime =
-            TestRuntime::make_runtime(TestSettings::new(), TestProvisioningResult::new(), ())
-                .wait()
-                .unwrap()
-                .with_module(Ok(module));
+        let runtime = TestRuntime::make_runtime(
+            TestSettings::new(),
+            TestProvisioningResult::new(),
+            TestHsm::default(),
+        )
+        .wait()
+        .unwrap()
+        .with_module(Ok(module));
         let handler = ListModules::new(runtime);
         let request = Request::get("http://localhost/modules")
             .body(Body::default())
@@ -174,11 +178,14 @@ mod tests {
     #[test]
     fn list_failed() {
         // arrange
-        let runtime =
-            TestRuntime::make_runtime(TestSettings::new(), TestProvisioningResult::new(), ())
-                .wait()
-                .unwrap()
-                .with_module(Err(Error::General));
+        let runtime = TestRuntime::make_runtime(
+            TestSettings::new(),
+            TestProvisioningResult::new(),
+            TestHsm::default(),
+        )
+        .wait()
+        .unwrap()
+        .with_module(Err(Error::General));
         let handler = ListModules::new(runtime);
         let request = Request::get("http://localhost/modules")
             .body(Body::default())
@@ -208,11 +215,14 @@ mod tests {
         // arrange
         let config = TestConfig::new("microsoft/test-image".to_string());
         let module = TestModule::new("test-module".to_string(), config, Err(Error::General));
-        let runtime =
-            TestRuntime::make_runtime(TestSettings::new(), TestProvisioningResult::new(), ())
-                .wait()
-                .unwrap()
-                .with_module(Ok(module));
+        let runtime = TestRuntime::make_runtime(
+            TestSettings::new(),
+            TestProvisioningResult::new(),
+            TestHsm::default(),
+        )
+        .wait()
+        .unwrap()
+        .with_module(Ok(module));
         let handler = ListModules::new(runtime);
         let request = Request::get("http://localhost/modules")
             .body(Body::default())
