+++ edgelet/edgelet-http-mgmt/src/server/module/start.rs
@@ -64,6 +64,7 @@ mod tests {
     use chrono::prelude::*;
     use edgelet_core::{MakeModuleRuntime, ModuleRuntimeState, ModuleStatus};
     use edgelet_http::route::Parameters;
+    use edgelet_test_utils::crypto::TestHsm;
     use edgelet_test_utils::module::*;
 
     use super::*;
@@ -82,11 +83,14 @@ mod tests {
         let config = TestConfig::new("microsoft/test-image".to_string());
         let module: TestModule<Error, _> =
             TestModule::new("test-module".to_string(), config, Ok(state));
-        let runtime =
-            TestRuntime::make_runtime(TestSettings::new(), TestProvisioningResult::new(), ())
-                .wait()
-                .unwrap()
-                .with_module(Ok(module));
+        let runtime = TestRuntime::make_runtime(
+            TestSettings::new(),
+            TestProvisioningResult::new(),
+            TestHsm::default(),
+        )
+        .wait()
+        .unwrap()
+        .with_module(Ok(module));
         let handler = StartModule::new(runtime);
         let parameters =
             Parameters::with_captures(vec![(Some("name".to_string()), "test".to_string())]);
@@ -114,11 +118,14 @@ mod tests {
         let config = TestConfig::new("microsoft/test-image".to_string());
         let module: TestModule<Error, _> =
             TestModule::new("test-module".to_string(), config, Ok(state));
-        let runtime =
-            TestRuntime::make_runtime(TestSettings::new(), TestProvisioningResult::new(), ())
-                .wait()
-                .unwrap()
-                .with_module(Ok(module));
+        let runtime = TestRuntime::make_runtime(
+            TestSettings::new(),
+            TestProvisioningResult::new(),
+            TestHsm::default(),
+        )
+        .wait()
+        .unwrap()
+        .with_module(Ok(module));
         let handler = StartModule::new(runtime);
         let request = Request::post("http://localhost/modules/test/start")
             .body(Body::default())
