+++ edgelet/edgelet-test-utils/src/module.rs
@@ -13,6 +13,8 @@ use futures::IntoFuture;
 use hyper::{Body, Request};
 use serde_derive::{Deserialize, Serialize};
 
+use crate::crypto::TestHsm;
+
 #[derive(Clone, Debug)]
 pub struct TestRegistry<E, C> {
     err: Option<E>,
@@ -264,7 +266,7 @@ where
     S::Config: Clone + Send + 'static,
 {
     type Config = S::Config;
-    type Crypto = ();
+    type Crypto = TestHsm;
     type Settings = S;
     type ProvisioningResult = TestProvisioningResult;
     type ModuleRuntime = Self;
